src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiohttp-cors==0.7.0
aiokafka==0.12.0
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorclass==2.2.2
colorlog==6.8.2
croniter==2.0.7
cryptography==43.0.3
fastapi==0.115.3
faust-streaming==0.11.3
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.10
intervaltree==3.1.0
kafka-python==1.4.7
lxml==5.3.0
mode==4.1.3
mode-streaming==0.4.1
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
numpy==2.1.3
opentracing==1.3.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
peewee==3.17.8
pip-tools==7.4.1
platformdirs==4.3.6
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
robinhood-aiokafka==1.1.6
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.0
terminaltables==3.1.10
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
venusian==3.1.0
webencodings==0.5.1
wheel==0.45.1
yarl==1.17.2
yfinance==0.2.49
redis==5.1.1
